requirements.txt
Django==1.10.3
Pillow==3.4.2
django-tinymce==2.4.0
django-widget-tweaks==1.3
easy-thumbnails==2.3
django-sessionprofile

# celery
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.20
kombu==3.0.33
django-celery-transactions2==0.1.6